Linter fixes - unconvert, revive:empty-lines, revive:var-naming, revive:unused-parameter

+++ plugins/inputs/kube_inventory/kube_state.go
@@ -166,7 +166,7 @@ func atoi(s string) int64 {
 	if err != nil {
 		return 0
 	}
-	return int64(i)
+	return i
 }
 
 func convertQuantity(s string, m float64) int64 {
